Export Projects API handlers
